Fixing some /w4 warnings by moving definitions to cpp file.

+++ include/vmath.h
@@ -1019,7 +1019,7 @@ static inline mat4 frustum(float left, float right, float bottom, float top, flo
 
 static inline mat4 perspective(float fovy, float aspect, float n, float f)
 {
-    float q = 1.0f / tan(radians(0.5f * fovy));
+    float q = 1.0f / (float)tan(radians(0.5f * fovy));
     float A = q / aspect;
     float B = (n + f) / (n - f);
     float C = (2.0f * n * f) / (n - f);
